fix: a string's length is a u32, and has been since P11.1-S1
`BigInt(binaryString)` segfaulted after `binaryString += "0"` ran
128 rounds. Sixteen call sites across twelve crates still read the
length as a u64 at offset 8 — the pre-P11.1-S1 shape. The field
narrowed to u32 back then and the four bytes above it became a
reserved slot; every one of those reads has been picking that slot
up along with the length ever since, and answering correctly only
because the slot was always zero.

The append kernel put a capacity there. So the reads started
answering `length | capacity << 32` — a number in the billions —
and the first consumer to trust it walked off the end of the heap.

Read the u32. Each site is the same one-line change, and the stale
`len: u64` comments that outlived the layout change go with it. The
`Tag::Arr` and dynobj reads at the same offset are untouched: those
fields really are u64.

The fixture builds a string by appending and hands it to every
reader in the list — BigInt from decimal and from a binary literal,
Map and Set keys, JSON.parse, Symbol description, Error message,
truthiness, concat. All byte-equal with bun; the BigInt line is the
one that used to exit 139.

+++ crates/torajs-anyvalue/src/arith.rs
@@ -367,11 +367,11 @@ unsafe fn add_operand_to_primitive(tag: i64, value: i64) -> (i64, i64, Option<An
 unsafe fn try_concat_short(l_str: *mut c_void, r_str: *mut c_void) -> Option<AnyValue> {
     let l_ptr = l_str as *const u8;
     let r_ptr = r_str as *const u8;
-    // SAFETY: Str layout invariant — `len: u64` at byte offset
-    // STR_LEN_OFF (= 8). Heap is 8-aligned so the u64 read is
-    // properly aligned.
-    let l_len = unsafe { (l_ptr.add(STR_LEN_OFF) as *const u64).read() } as usize;
-    let r_len = unsafe { (r_ptr.add(STR_LEN_OFF) as *const u64).read() } as usize;
+    // SAFETY: Str layout invariant — `length: u32` at byte offset
+    // STR_LEN_OFF (= 8), with the capacity slot in the four bytes
+    // above it.
+    let l_len = unsafe { (l_ptr.add(STR_LEN_OFF) as *const u32).read() } as usize;
+    let r_len = unsafe { (r_ptr.add(STR_LEN_OFF) as *const u32).read() } as usize;
     let total = l_len + r_len;
     if total > SHORT_STR_CAP {
         return None;
